src/requirements.txt
pydantic
fastapi
starlette
uvicorn
pymongo
python-dotenv
httpx
aiofiles
azure-storage-blob